requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
deep-translator==1.11.4
idna==3.10
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
reportlab==4.2.5
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
